chore(bingads): move sdk loader to separate file
